server_2_for_display/python_backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
Pillow
python-dotenv
ultralytics
opencv-python
